Add files via upload
Hi, this is simon says game. In this game we have four colored buttons. The game starts with press any key and then lighting one or more buttons in a random sequence. players then need to repeat the sequence by pressing the buttons in the same order. The sequence get longer and complex as the game progresses. if you press wrong button you lose the game.